requirements.txt
Pillow
mss
numpy
opencv-python
matplotlib
pyautogui
screeninfo
pywinauto